requirements.txt
dearpygui==2.1.0
importlib_metadata==8.7.0
packaging==25.0
pnio_dcp==1.2.0
psutil==7.0.0
pyparsing==3.2.3
python-snap7==2.0.2
setuptools==80.9.0
setuptools-scm==8.3.1
zipp==3.23.0
